+++ src/assets/css/header.css
@@ -1,0 +1,162 @@
+:root {
+  /* RoyalBlue */
+  --header-text-color: #4169e1;
+}
+
+/* 상단 바 */
+#main_header {
+  width: 960px;
+  margin: 0 auto;
+
+  height: 160px;
+  position: relative;
+}
+
+#main_header > nav > * {
+  top: 50px;
+}
+
+#title {
+  position: absolute;
+  left: 20px;
+  font-size: 40px;
+}
+
+#menu {
+  position: absolute;
+  right: 0;
+  font-size: 24px;
+}
+
+/* 타이틀 */
+#main_header {
+  font-family: "Archivo", sans-serif;
+  font-family: "Coiny", cursive;
+}
+
+#main_header a {
+  color: var(--header-text-color);
+  text-decoration: none;
+}
+
+/* login menu */
+#menu > ul > span {
+  overflow: hidden;
+  float: left;
+}
+
+#menu > ul > span > li {
+  float: left;
+}
+
+#menu a {
+  padding: 2px 10px;
+}
+
+.hidden {
+  display: none;
+}
+
+/* :root {
+  --text-color: #f0f4f5;
+  --background-color: #263343;
+  --accent-color: pink;
+}
+
+body {
+  margin: 0;
+  font-family: "Source Sans Pro";
+}
+
+a {
+  text-decoration: none;
+  color: var(--text-color);
+}
+
+.navbar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background-color: var(--background-color);
+  padding: 8px 12px;
+}
+
+.navbar_logo {
+  font-size: 24px;
+  color: var(--text-color);
+}
+
+.navbar_logo i {
+  color: var(--accent-color);
+}
+
+.navbar_menu {
+  display: flex;
+  list-style: none;
+  padding-left: 0px;
+}
+
+.navbar_menu li {
+  padding: 8px 12px;
+}
+
+.navbar_menu li:hover {
+  background-color: var(--accent-color);
+  border-radius: 4px;
+}
+
+.navbar_icons {
+  list-style: none;
+  color: var(--text-color);
+  display: flex;
+  padding-left: 0px;
+}
+
+.navbar_icons li {
+  padding: 8px 12px;
+}
+
+.navbar_toggleBtn {
+  display: none;
+  position: absolute;
+  right: 32px;
+  font-size: 24px;
+  color: var(--accent-color);
+}
+
+
+
+@media screen and (max-width: 768px) {
+  .navbar {
+    flex-direction: column;
+    align-items: flex-start;
+    padding: 8px 24px;
+  }
+
+  .navbar_menu {
+    display: none;
+    flex-direction: column;
+    align-items: center;
+    width: 100%;
+  }
+
+  .navbar_menu li {
+    width: 100%;
+    text-align: center;
+  }
+
+  .navbar_icons {
+    display: none;
+    justify-content: center;
+    width: 100%;
+  }
+
+  .navbar_toggleBtn {
+    display: block;
+  }
+
+  .navbar_menu.active,
+  .navbar_icons.active {
+    display: flex;
+  }
+} */
